Added Sliding
The player is now able to slide/crouch around the map. Also updated some testing scenes to playtest. The crouching behaviour gets handled in the CrouchMovement script. Some of the values within the main movement script needed to be changed while sliding. So I created some getters and setters for the needed values. This way almost everything stays loose and private. lastly I started adding some particles to the player. This way the player receives visual feedback for their speed or spped on impact with the ground/walls.