Added wallrunning
Added wallrunning mechanic. The wallrunning is handled in the WallrunMovement script. This works by checking if there are walls close to the player. The wall-checking works via raycasts, these raycasts look for a detection on the whatIsWall layer. If the player is next to a wall and is giving the correct inpit, he or she will start wallrunning. This works by toggeling the gravity on the rigidbody and applying forces to the player, so that the player sticks to the wall. The player can also jump from the wall when pressing space. Lastly I implemented the audio scripts from my previous project. I can use this audiomanager to easily play sound effects. This works by using the instance method. This way everything stays nice and loose from eachother!