Added item pickups and made prefabs
Added a weapon pickup script. This script checks if there is a weapon in range, based on a raycast. This raycast checks for weapons, when it triggers one the player can choose to pick it up. The player can also drop the weapon whenever they want to.
The weapon also gets moved to a specific weapon layer, so that the weapons can be renderer above the map. This prevents the weapon models from going through walls! Lastly I made some prefabs for the weapon system, and added some materials.