Added a custom TimeManager
This script can be used to slow down the time of the game. This opens op some awesome slowmo moments. I also stated with some the enemy AI. I found some animations from Mixamo, and imported them to the models I wanted to use. Lastly I created some prefabs for the enemy. This way I only need to create the scripts and put them on the prefab!