Added AI scripts
Added the AI behaviour. The AI chooses a dynamically chosen spot to take cover. This spot get chosen based on a couple of values. The main values are the colliders that it takes cover behind. The AI finds these colliders and takes cover behind them. After taking cover the AI begins to shoot if the player is nearby and in sight.
The enemy also has some sound and visual effects. If the player collides with the bullet, the player takes damage.