Implemented Dialogue system
Added the system and some audiofiles (dialogues)for the game. The system checks if the player triggered an dialogue. If the player did, the system checks for dialogues. The system plays the array/dialogues that need to be said. If the sentence is complete, and the audioclip is also done playing. The system continues with the next system in the que. I implemented this system, because I wanted to guide the player's that have some trouble while playing the game.